LayoutTests/fast/events/element-click-no-coords-expected.txt

Tests that the MouseEvent's coordinates are 0 for a simulated click.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.screenX is 0
PASS event.screenY is 0
PASS event.clientX is 0
PASS event.clientY is 0
PASS event.pageX is 0
PASS event.pageY is 0
PASS successfullyParsed is true

TEST COMPLETE
Test Div
